Unirent_App/requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
pymongo==4.1.1
apscheduler==3.9.1
python-dotenv==0.19.2
requests==2.28.1